pulled functionality out into unique files
