+++ lib/verify-stream.js
@@ -1,0 +1,120 @@
+/*global module*/
+const base64url = require('base64url');
+const DataStream = require('./data-stream');
+const jwa = require('jwa');
+const Stream = require('stream');
+const toString = require('./tostring');
+const util = require('util');
+const JWS_REGEX = /^[a-zA-Z0-9\-_]+?\.[a-zA-Z0-9\-_]+?\.([a-zA-Z0-9\-_]+)?$/;
+
+function isObject(thing) {
+  return Object.prototype.toString.call(thing) === '[object Object]';
+}
+
+function safeJsonParse(thing) {
+  if (isObject(thing))
+    return thing;
+  try { return JSON.parse(thing); }
+  catch (e) { return undefined; }
+}
+
+function headerFromJWS(jwsSig) {
+  const encodedHeader = jwsSig.split('.', 1)[0];
+  return safeJsonParse(base64url.decode(encodedHeader, 'binary'));
+}
+
+function securedInputFromJWS(jwsSig) {
+  return jwsSig.split('.', 2).join('.');
+}
+
+function algoFromJWS(jwsSig) {
+  var err;
+  const header = headerFromJWS(jwsSig);
+  if (typeof header != 'object') {
+    err = new Error("Invalid token: no header in signature '" + jwsSig + "'");
+    err.code = "MISSING_HEADER";
+    err.signature = jwsSig;
+    throw err;
+  }
+  if (!header.alg) {
+    err = new Error("Missing `alg` field in header for signature '"+ jwsSig +"'");
+    err.code = "MISSING_ALGORITHM";
+    err.header = header;
+    err.signature = jwsSig;
+    throw err;
+  }
+  return header.alg;
+}
+
+function signatureFromJWS(jwsSig) {
+  return jwsSig.split('.')[2];
+}
+
+function payloadFromJWS(jwsSig) {
+  const payload = jwsSig.split('.')[1];
+  return base64url.decode(payload, 'binary');
+}
+
+function isValidJws(string) {
+  return JWS_REGEX.test(string) && !!headerFromJWS(string);
+}
+
+function jwsVerify(jwsSig, secretOrKey) {
+  jwsSig = toString(jwsSig);
+  const signature = signatureFromJWS(jwsSig);
+  const securedInput = securedInputFromJWS(jwsSig);
+  const algo = jwa(algoFromJWS(jwsSig));
+  return algo.verify(securedInput, signature, secretOrKey);
+}
+
+function jwsDecode(jwsSig, opts) {
+  opts = opts || {};
+  jwsSig = toString(jwsSig);
+  if (!isValidJws(jwsSig))
+    return null;
+  const header = headerFromJWS(jwsSig);
+  if (!header)
+    return null;
+  var payload = payloadFromJWS(jwsSig);
+  if (header.typ === 'JWT' || opts.json)
+    payload = JSON.parse(payload);
+  return {
+    header: header,
+    payload: payload,
+    signature: signatureFromJWS(jwsSig)
+  };
+}
+
+function VerifyStream(opts) {
+  opts = opts || {};
+  const secretOrKey = opts.secret||opts.publicKey||opts.key;
+  const secretStream = new DataStream(secretOrKey);
+  this.readable = true;
+  this.secret = this.publicKey = this.key = secretStream;
+  this.signature = new DataStream(opts.signature);
+  this.secret.once('close', function () {
+    if (!this.signature.writable && this.readable)
+      this.verify();
+  }.bind(this));
+
+  this.signature.once('close', function () {
+    if (!this.secret.writable && this.readable)
+      this.verify();
+  }.bind(this));
+}
+util.inherits(VerifyStream, Stream);
+VerifyStream.prototype.verify = function verify() {
+  const valid = jwsVerify(this.signature.buffer, this.key.buffer);
+  const obj = jwsDecode(this.signature.buffer);
+  this.emit('done', valid, obj);
+  this.emit('data', valid);
+  this.emit('end');
+  this.readable = false;
+  return valid;
+};
+
+VerifyStream.decode = jwsDecode;
+VerifyStream.isValid = isValidJws;
+VerifyStream.verify = jwsVerify;
+
+module.exports = VerifyStream;
